Refactor: Extract transition sweep effect into a utility function
This commit extracts the drawing logic for the transition sweep effect in `BoardPreview` into a new utility function `drawTransitionSweepEffect` within `DrawUtils.kt`.

**Key Changes:**

- **`DrawUtils.kt`:**
    - Added a new internal extension function `DrawScope.drawTransitionSweepEffect`.
    - This function encapsulates the logic for drawing a horizontal gradient sweep and a vertical line based on the provided `progress`, `isRunning` state, `color`, and `maxWidth`.
- **`BoardPreview.kt`:**
    - Replaced the inline drawing logic for the transition sweep effect with a call to the new `drawTransitionSweepEffect` utility function.
    - The parameters `progress`, `state.isRunning`, `colors.frame`, and `maxWidth` are passed to the new function.

This refactoring improves code organization and reusability by separating the specific drawing logic into a dedicated, testable function.
