+++ feature/creator/src/main/java/com/example/feature/creator/components/preview/BoardPreview.kt
@@ -8,8 +8,6 @@ import androidx.compose.ui.draw.drawWithContent
 import androidx.compose.ui.geometry.CornerRadius
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.drawscope.clipRect
 import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
@@ -80,36 +78,12 @@ internal fun BoardPreview(
 				)
 			}
 
-			if (state.isRunning) {
-				val glowWidth = 20.dp.toPx()
-
-				val sweepBrush = Brush.horizontalGradient(
-					colorStops = arrayOf(
-						0.0f to Color.Transparent,
-						0.3f to colors.frame.copy(alpha = 0.05f),
-						0.6f to colors.frame.copy(alpha = 0.2f),
-						0.8f to colors.frame.copy(alpha = 0.4f),
-						0.9f to colors.frame.copy(alpha = 0.2f),
-						1.0f to Color.Transparent,
-					),
-					startX = (maxWidth * progress) - glowWidth,
-					endX = (maxWidth * progress) + (glowWidth * 0.6f),
-				)
-
-				drawRect(
-					brush = sweepBrush,
-					topLeft = Offset(x = 0f, y = 0f),
-					size = Size(width = maxWidth, height = size.height),
-				)
-
-				drawLine(
-					color = colors.frame.copy(alpha = 0.7f),
-					start = Offset(x = maxWidth * progress, y = 0f),
-					end = Offset(x = maxWidth * progress, y = size.height),
-					strokeWidth = 1.dp.toPx(),
-					alpha = 0.5f,
-				)
-			}
+			drawTransitionSweepEffect(
+				progress = progress,
+				isRunning = state.isRunning,
+				color = colors.frame,
+				maxWidth = maxWidth,
+			)
 		},
 	)
 }
